Moved cv::Mat serialisation stuff to utils/ subdirectory
