+++ CMakeLists.txt
@@ -45,8 +45,8 @@ set(CEREAL_INCLUDE_DIR "${CMAKE_SOURCE_DIR}/3rdparty/cereal-1.1.1/include")
 set(HEADERS
 	include/superviseddescent/superviseddescent.hpp
 	include/superviseddescent/regressors.hpp
-	include/superviseddescent/matserialisation.hpp
-	include/superviseddescent/matcerealisation.hpp
+	include/superviseddescent/utils/mat_cerealisation.hpp
+	include/superviseddescent/utils/mat_serialization.hpp
 	include/superviseddescent/utils/ThreadPool.h
 )
