+++ include/superviseddescent/utils/mat_cerealisation.hpp
